+++ openflow-protocol-impl/src/test/java/org/opendaylight/openflowjava/protocol/impl/deserialization/factories/PacketInMessageFactoryTest.java
@@ -29,7 +29,7 @@ public class PacketInMessageFactoryTest {
         Assert.assertEquals("Wrong bufferID", 0x00010203L, builtByFactory.getBufferId().longValue());
         Assert.assertEquals("Wrong totalLength", 0x0102, builtByFactory.getTotalLen().intValue());
         Assert.assertEquals("Wrong reason", 0x01, builtByFactory.getReason().shortValue());
-        Assert.assertEquals("Wrong tableID", new TableId((long) 4), builtByFactory.getTableId());
+        Assert.assertEquals("Wrong tableID", new TableId(4L), builtByFactory.getTableId());
         Assert.assertEquals("Wrong cookie", 0x0001020304050607L, builtByFactory.getCookie().longValue());
         Assert.assertArrayEquals("Wrong data", ByteBufUtils.hexStringToBytes("01 02 03 04"), builtByFactory.getData());
     }
